Rename mip._variable module to mip._variables.
