+++ ocean/mip/__init__.py
@@ -1,7 +1,7 @@
 from ._base import BaseModel
 from ._explanation import MixedIntegerProgramExplanation
 from ._model import Model
-from ._variable import FeatureVar, TreeVar
+from ._variables import FeatureVar, TreeVar
 
 __all__ = [
     "BaseModel",
